% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R
\name{causal_forest_importance_scores}
\alias{causal_forest_importance_scores}
\title{Internal function called to return the importance scores from causal forest}
\usage{
causal_forest_importance_scores(
  X,
  y,
  trt,
  type = "regression",
  shuffle = FALSE,
  ...
)
}
\arguments{
\item{X}{original data.frame with "numeric" and "factor" columns only.}

\item{y}{response vector with \code{length(y) = nrow(X)}. Accepts "numeric" (family="gaussian") or binary "factor" (family="binomial"). Can also be a survival object of class Surv
as obtained from y = survival::Surv(time, status).}

\item{trt}{binary treatment indicator}

\item{type}{should be "regression" if y is numeric, "classification" if y is a binary factor variable or "survival" if y is a survival object.}

\item{shuffle}{boolean variable, that takes the value FALSE if we do not want the target y to be shuffled, and TRUE otherwise}

\item{...}{}
}
\value{
importance scores
}
\description{
Internal function called to return the importance scores from causal forest
}
\keyword{internal}
