% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R
\name{check_normality}
\alias{check_normality}
\title{Internal check of normality of the numeric input covariates}
\usage{
check_normality(X)
}
\arguments{
\item{X}{data frame or tibble}
}
\value{
this function doesn't return anything
}
\description{
Do not call this function on its own
}
\keyword{internal}
