% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/performance.R
\name{eval_tpp}
\alias{eval_tpp}
\title{True positive proportion (tpp) as function of selection and known positives:}
\usage{
eval_tpp(selected, positives)
}
\arguments{
\item{selected}{vector of indices of selected variables}

\item{positives}{vector of indices of known non-null variables (that influence response)}
}
\value{
true positive rate
}
\description{
True positive proportion (tpp) as function of selection and known positives:
}
\examples{
#' library(knockofftools)

eval_tpp(selected=c(1,2,3,5), positives=1:4)
}
