% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R
\name{find_single_optimal_variable_set}
\alias{find_single_optimal_variable_set}
\title{Select variables based on (heuristic) mode of multiple variable selections}
\usage{
find_single_optimal_variable_set(S, p, trim = 0.5)
}
\arguments{
\item{S}{list of variable selection indices}

\item{p}{number of variables. Each element of the list of selection indices should be a subset of 1:p.}

\item{trim}{trimming probability threshold. A sensible default is \code{trim=0.5}.}
}
\value{
a single "most frequent" variable selection among the multiple selections in S.
}
\description{
Do not call this function on its own
}
\keyword{internal}
