% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simdata.R
\name{generate_lp}
\alias{generate_lp}
\title{Generate linear predictor with first p_nn beta coefficients = a, all other = 0}
\usage{
generate_lp(X, p_nn, a)
}
\arguments{
\item{X}{data.frame with numeric and factor columns only}

\item{p_nn}{number of non-null covariate predictors.
The regression coefficients (beta) corresponding to
columns 1:p_nn of x will be non-zero, all other are set to zero.}

\item{a}{amplitude of non-null regression coefficients}
}
\value{
linear predictor X%*%beta
}
\description{
Generate linear predictor with first p_nn beta coefficients = a, all other = 0
}
\keyword{internal}
