% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simdata.R
\name{generate_simdata}
\alias{generate_simdata}
\title{Generate simulated data set for vignettes}
\usage{
generate_simdata()
}
\value{
generates the simulated data as obtained by the call data(simdata)
}
\description{
Generate simulated data set for vignettes
}
\examples{
library(knockofftools)
data(simdata)
all.equal(generate_simdata(), simdata)
}
