% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R
\name{glasso_adjacency_matrix}
\alias{glasso_adjacency_matrix}
\title{Estimate adjacency matrix using graphical LASSO (glasso)}
\usage{
glasso_adjacency_matrix(X)
}
\arguments{
\item{X}{data.frame (or tibble) of covariates}
}
\value{
adjacency matrix
}
\description{
Estimate adjacency matrix using graphical LASSO (glasso)
}
