% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simknockoffs.R
\name{knockoff}
\alias{knockoff}
\title{Knockoffs for general covariate data frames}
\usage{
knockoff(X, method = "seq", ...)
}
\arguments{
\item{X}{data.frame (or tibble) with "numeric" and "factor" columns only. The number of columns, ncol(X) needs to be > 2.}

\item{method}{what type of knockoffs to calculate. Defaults to sequential knockoffs, method="seq", which works for
both numeric and factor variables. The other options are method="sparseseq", which is the sparse sequential knockoff generation algorithm, and method="mx", which only works if all columns of X are continuous.}

\item{...}{additional parameters passed to the downstream knockoffs generating functions}
}
\value{
knockoff copy of X
}
\description{
Knockoffs for general covariate data frames
}
\examples{
library(knockofftools)

set.seed(1)

X <- generate_X(n=100, p=6, p_b=2, cov_type="cov_equi", rho=0.5)

# sequential knockoff:
Xk <- knockoff(X)

X <- generate_X(n=100, p=6, p_b=0, cov_type="cov_equi", rho=0.5)

# MX-knockoff:
Xk <- knockoff(X, method="mx")
}
