% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R
\name{ns.transform}
\alias{ns.transform}
\title{Normal score transformation function}
\usage{
ns.transform(y)
}
\arguments{
\item{y}{a numeric vector representing the continuous variable}
}
\value{
a vector of length(y) with the normal-score transformed variable
}
\description{
Normal score transformation function
}
\keyword{internal}
