% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R
\name{ranger_arg_checks}
\alias{ranger_arg_checks}
\title{Internal function called in ranger_importance_scores to check and
modify arguments for ranger to be suitable for knockoff statistic computation}
\usage{
ranger_arg_checks(dots)
}
\arguments{
\item{dots}{A list of arguments to be passed to \code{ranger::ranger}.}
}
\value{
A modified list of arguments with valid importance and mtry settings.
}
\description{
Internal function called in ranger_importance_scores to check and
modify arguments for ranger to be suitable for knockoff statistic computation
}
\keyword{internal}
