% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R
\name{ranger_importance_scores}
\alias{ranger_importance_scores}
\title{Internal function called to return the importance scores from random forest
(ranger)}
\usage{
ranger_importance_scores(
  X,
  y,
  type = c("regression", "classification", "survival"),
  max_retry = 2,
  ...
)
}
\arguments{
\item{X}{original data.frame with "numeric" and "factor" columns only.}

\item{y}{response vector with \code{length(y) = nrow(X)}. Accepts "numeric" (family="gaussian") or binary "factor" (family="binomial"). Can also be a survival object of class Surv
as obtained from y = survival::Surv(time, status).}

\item{type}{should be "regression" if y is numeric, "classification" if y is a binary factor variable or "survival" if y is a survival object.}

\item{max_retry}{number of times to retry ranger if NaN importance scores are reported. Default is 2.}

\item{...}{additional arguments passed to \code{ranger::ranger}}
}
\value{
importance scores
}
\description{
Internal function called to return the importance scores from random forest
(ranger)
}
\author{
Maike Ahrens, Sebastian Voss
}
\keyword{internal}
