% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knockoff_filters.R
\name{selections_control_FDR}
\alias{selections_control_FDR}
\title{Controls the false discovery rate (FDR) given knockoff W-statistics.}
\usage{
selections_control_FDR(W, level)
}
\arguments{
\item{W}{a vector of knockoff W-statistics (feature statistics).}

\item{level}{The nominal level at which to control the FDR.}
}
\value{
the selected variables
}
\description{
This method was introduced by Candes et al. (2018)
}
\details{
E. Candes, Y. Fan, L. Janson, & J. Lv, (2018). Panning for gold:‘model‐X’knockoffs for high dimensional controlled variable selection. Journal of the Royal Statistical Society: Series B (Statistical Methodology), 80(3), 551-577.
}
\examples{
W <- rnorm(100)
selections_control_FDR(W, level=0.5)

}
