% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knockoff_filters.R
\name{selections_control_kFWER}
\alias{selections_control_kFWER}
\title{Controls the k-familywise error rate (k-FWER) given a vector of knockoff W-statistics.}
\usage{
selections_control_kFWER(W, level, k)
}
\arguments{
\item{W}{a vector of knockoff W-statistics (feature statistics)}

\item{level}{The nominal level at which to control the k-FWER.}

\item{k}{a positive integer corresponding to k-FWER (multiple testing when one seeks to control at least k false discoveries)}
}
\value{
the selected variables
}
\description{
This method was introduced by Janson and Su (2016),, while we used the implementation https://github.com/zhimeir/derandomized_knockoffs_paper/blob/master/R/vanilla_fwer_filter.R
}
\details{
L. Janson, & W. Su, (2016). Familywise error rate control via knockoffs. Electronic Journal of Statistics, 10(1), 960-975.
}
\examples{
W <- rnorm(100)
selections_control_kFWER(W, level=0.1, k=5)

}
