% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knockoff_filters.R
\name{selections_control_PFER}
\alias{selections_control_PFER}
\title{Controls the per-familywise error rate (PFER) given knockoff W-statistics.}
\usage{
selections_control_PFER(W, level)
}
\arguments{
\item{W}{a vector of knockoff W-statistics (feature statistics)}

\item{level}{The nominal level at which to control the PFER.}
}
\value{
the selected variables
}
\description{
This method was introduced by Janson and Su (2016), while we used the implementation of https://github.com/zhimeir/derandomized_knockoffs_paper/blob/master/R/pfer_filter.R
}
\details{
L. Janson, & W. Su, (2016). Familywise error rate control via knockoffs. Electronic Journal of Statistics, 10(1), 960-975.
}
\examples{
W <- rnorm(100)
selections_control_PFER(W, level=2)

}
