% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simknockoffs.R
\name{sim_EN}
\alias{sim_EN}
\title{Simulate from elastic-net regression model}
\usage{
sim_EN(y, X, ...)
}
\arguments{
\item{y}{response vector (either "numeric" or "factor") that gets passed to cv.glmnet}

\item{X}{data.frame of covariates that are passed to cv.glmnet}

\item{...}{other parameters passed to the function cv.glmnet}
}
\value{
simulated response
}
\description{
Simulate from elastic-net regression model
}
\examples{
library(knockofftools)

set.seed(1)

X = data.frame(matrix(rnorm(100 * 20), 100, 20))
y = X[,1] + rnorm(100)

# simulate from elastic-net regression:
ysim = sim_EN(y=y, X=X)

# simulated versus input response:
plot(y, ysim)
}
