% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simknockoffs.R
\name{sim_simple}
\alias{sim_simple}
\title{Simple knockoff generator based on least squares fit (continuous variables) or multinomial logistic regression (factor variables) respectively.
If X is empty, knockoffs are sampled from the marginal distribution of y}
\usage{
sim_simple(y, X)
}
\arguments{
\item{y}{response vector (either "numeric" or "factor") that gets passed to cv.glmnet}

\item{X}{data.frame of covariates that are passed to cv.glmnet}
}
\value{
simulated response
}
\description{
Simple knockoff generator based on least squares fit (continuous variables) or multinomial logistic regression (factor variables) respectively.
If X is empty, knockoffs are sampled from the marginal distribution of y
}
