% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simdata-data.R
\docType{data}
\name{simdata}
\alias{simdata}
\title{Simulated dataset for knockofftools}
\format{
A data frame with 2000 rows and 33 variables:
\describe{
  \item{Tc}{Treatment indicator}
  \item{X1--X30}{Covariates}
  \item{Yb}{Binary outcome}
  \item{Yg}{Continuous outcome}
}
}
\source{
Simulated using \code{generate_simdata()}
}
\usage{
simdata
}
\description{
A synthetic dataset generated by the function \code{generate_simdata()}.
It contains simulated Gaussian, binary, and survival outcomes along with covariates.
}
\examples{
data(simdata)
head(simdata)
}
\keyword{datasets}
