% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knockoff_filters.R
\name{stat_custom}
\alias{stat_custom}
\title{Construction helper for knockoff (feature) statistics}
\usage{
stat_custom(
  X,
  X_k,
  y,
  type = c("regression", "classification", "survival"),
  model_imp,
  ...
)
}
\arguments{
\item{X}{original data.frame (or tibble) with "numeric" and "factor" columns only. The number of columns, ncol(X) needs to be > 2.}

\item{X_k}{knockoff data.frame (or tibble) with "numeric" and "factor" columns only obtained e.g. by X_k = knockoff(X). The dimensions and column classes must match
those of the original X.}

\item{y}{response vector with \code{length(y) = nrow(X)}. Accepts "numeric" (family="gaussian") or binary "factor" (family="binomial"). Can also be a survival object of class "Surv" (type="survival")
as obtained from y = survival::Surv(time, status).}

\item{type}{should be "regression" if y is numeric, "classification" if y is a binary factor variable or "survival" if y is a survival object.}

\item{model_imp}{function that takes feature input X, the outcome y,
outcome type (regression, classification, survival), and any additional
parameters passed via `...`, fits a model, computes and returns
a vector of variable importance scores.}

\item{...}{Other arguments to be passed to the underlying modeling or
importance function from `model_imp`.}
}
\value{
data.frame with knockoff statistics W as column.
The number of rows matches the number of columns (variables)
of the data.frame X and the variable names are recorded in rownames(W).
}
\description{
The `stat_custom` function is a helper function that allows
users to define their own knockoff feature statistics by simply
providing a custom modeling and importance function.
`stat_custom` takes care of swapping the original and knockoff variables
before the custom computations which avoids bias introduction in the
the knockoff statistics and corrects the sign of the computed statistics
accordingly before returning them.
}
