% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knockoff_filters.R
\name{stat_ranger}
\alias{stat_ranger}
\title{Knockoff (feature) statistics: Random forest (ranger)}
\usage{
stat_ranger(X, X_k, y, type, ...)
}
\arguments{
\item{X}{original data.frame (or tibble) with "numeric" and "factor" columns only. The number of columns, ncol(X) needs to be > 2.}

\item{X_k}{knockoff data.frame (or tibble) with "numeric" and "factor" columns only obtained e.g. by X_k = knockoff(X). The dimensions and column classes must match
those of the original X.}

\item{y}{response vector with \code{length(y) = nrow(X)}. Accepts "numeric" (family="gaussian") or binary "factor" (family="binomial"). Can also be a survival object of class "Surv" (type="survival")
as obtained from y = survival::Surv(time, status).}

\item{type}{should be "regression" if y is numeric, "classification" if y is a binary factor variable or "survival" if y is a survival object.}

\item{...}{other parameters passed to \code{ranger::ranger()}
controlling modelling and the importance metric.}
}
\value{
data.frame with knockoff statistics W as column.
The number of rows matches the number of columns (variables) of the
data.frame X and the variable names are recorded in `rownames(W)`.
}
\description{
This function uses the `{ranger}` package to estimate
permutation importance scores from random forest (ranger).
}
\seealso{
\code{\link[knockofftools]{stat_random_forest}}
}
