% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R
\name{check_if_continuous}
\alias{check_if_continuous}
\title{Heuristic check for whether a variable can be reasonably treated as continuous}
\usage{
check_if_continuous(X)
}
\arguments{
\item{x}{a numeric variable vector}
}
\value{
a logical TRUE or FALSE depending on whether n_distinct(x) > 30
}
\description{
Heuristic check for whether a variable can be reasonably treated as continuous
}
\keyword{internal}
