% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knockoff_filters.R
\name{.knockoff.statistics.single}
\alias{.knockoff.statistics.single}
\title{Knockoff (feature) statistics for a single knockoff}
\usage{
.knockoff.statistics.single(
  y,
  X,
  type = "regression",
  knockoff.method = "seq",
  statistic = "stat_glmnet",
  trt = NULL,
  adjacency.matrix = NULL,
  ...
)
}
\arguments{
\item{y}{response vector with \code{length(y) = nrow(X)}. Accepts "numeric", binary "factor", or survival ("Surv") object.}

\item{X}{data.frame (or tibble) with "numeric" and "factor" columns only. The number of columns, ncol(X) needs to be > 2.}

\item{type}{should be "regression" if y is numeric, "classification" if y is a binary factor variable or "survival" if y is a survival object.}

\item{knockoff.method}{what type of knockoffs to calculate. Defaults to sequential knockoffs, knockoff.method="seq", but other options are "sparseseq" and "mx". The "mx" option only works if all columns of X are continuous.}

\item{statistic}{knockoff feature statistic function, defaults to glmnet coefficient difference (statistic="stat_glmnet"; see ?stat_glmnet). Other options include statistic="stat_random_forest" (see ?stat_random_forest), statistic="stat_predictive_glmnet" (see ?stat_predictive_glmnet) or statistic="stat_predictive_causal_forest" (see ?stat_predictive_causal_forest).}

\item{trt}{binary treatment (factor) variable required if statistic involves a predictive knockoff filter (i.e. if statistic="stat_predictive_glmnet" or statistic="stat_predictive_causal_forest")}

\item{adjacency.matrix}{optional user specified adjacency matrix (i.e. binary indicator matrix corresponding to the non-zero elements of the precision matrix of X). Defaults to NULL and is then estimated within the function call.}

\item{...}{additional parameters passed to the "statistic" function (note that the knockoffs parameter X_k should not be entered by user; it is already calculated inside the knockoff.statistics function).}

\item{M}{the number of independent knockoff feature statistics that should be calculated.}
}
\value{
data.frame with a single knockoff statistics W as column.
}
\description{
Do not call this function on its own. Use \code{knockoff.statistics} instead.
}
\keyword{internal}
