% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knockoff_filters.R
\name{multi_select}
\alias{multi_select}
\title{Select variables based on the heuristic multiple selection algorithm from Kormaksson et al. 'Sequential
knockoffs for continuous and categorical predictors: With application to a large psoriatic arthritis clinical
trial pool.' Statistics in Medicine. 2021;1–16.}
\usage{
multi_select(S, trim = 0.5)
}
\arguments{
\item{S}{the binary matrix of selections}

\item{trim}{trimming probability threshold. A sensible default is \code{trim=0.5}.}
}
\value{
a single "most frequent" variable selection among the multiple selections in S.
}
\description{
Select variables based on the heuristic multiple selection algorithm from Kormaksson et al. 'Sequential
knockoffs for continuous and categorical predictors: With application to a large psoriatic arthritis clinical
trial pool.' Statistics in Medicine. 2021;1–16.
}
\details{
M. Kormaksson, L. J. Kelly, X. Zhu, S. Haemmerle, L. Pricop, & D. Ohlssen (2021). Sequential knockoffs for continuous and categorical predictors: With application to a large psoriatic arthritis clinical trial pool. Statistics in Medicine, 40(14), 3313-3328.
}
\examples{
library(knockofftools)

set.seed(1)

p = 31
Nknockoff = 100
S <- matrix(sample(0:1,p*Nknockoff, replace=TRUE), p, Nknockoff)

multi_select(S)
}
