% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R
\name{random_forest_importance_scores}
\alias{random_forest_importance_scores}
\title{Internal function called to return the importance scores from random forest}
\usage{
random_forest_importance_scores(X, y, trt, type = "regression", ...)
}
\arguments{
\item{X}{original data.frame with "numeric" and "factor" columns only.}

\item{y}{response vector with \code{length(y) = nrow(X)}. Accepts "numeric" (family="gaussian") or binary "factor" (family="binomial"). Can also be a survival object of class Surv
as obtained from y = survival::Surv(time, status).}

\item{type}{should be "regression" if y is numeric, "classification" if y is a binary factor variable or "survival" if y is a survival object.}

\item{...}{}
}
\value{
importance scores
}
\description{
Internal function called to return the importance scores from random forest
}
\keyword{internal}
